Feat: Raw query interface and index file refactor
